feat(coveralls): basic implementation of plugin
I wrote this because grunt-karma-coveralls is pretty cool, but I don't very much care to use grunt.
This is going to be buggy for a while, but I'm hoping this can sort of turn into a "good enough"
sort of thing, for people testing web applications using Karma and Istanbul.