Encourage people to contribute even harder
The whole "I didn't send a PR because I'm not sure how you'd want to handle it" thing is weird, folks. I'm not your bosses, I built this thing for my own uses, and to give something back to the JS community, who could hopefully find it useful --- and if it needs something that it doesn't currently have, and you know how to add it, please do. DO IT!